feat: modificar parâmetros de entrada e retornos para os respectivos DTOs

+++ src/main/java/tech/ada/java/agendamentoconsultas/service/DoctorService.java
@@ -1,42 +1,63 @@
 package tech.ada.java.agendamentoconsultas.service;
 
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import tech.ada.java.agendamentoconsultas.exception.DoctorNotFoundException;
 import tech.ada.java.agendamentoconsultas.model.Doctor;
+import tech.ada.java.agendamentoconsultas.model.Dto.DoctorDtoRequest;
+import tech.ada.java.agendamentoconsultas.model.Dto.DoctorDtoResponse;
 import tech.ada.java.agendamentoconsultas.respository.DoctorRepository;
 
 import java.util.List;
 import java.util.UUID;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 
 public class DoctorService {
     private final DoctorRepository repository;
-    public List<Doctor> findALl() {
+    private final ModelMapper modelMapper;
 
-        return this.repository.findAll();
-    };
+    public List<DoctorDtoResponse> findALl() {
 
-    public Doctor findByUuid(UUID uuid){
-      return this.repository.findByUuid(uuid).orElseThrow(RuntimeException::new);
-    };
-
-    public Doctor addDoctor(Doctor doctor){
-      return repository.save(doctor);
-    };
-
-    public void update(UUID uuid, Doctor newDoctor){
-        Doctor doctor = findByUuid(uuid);
-
-        newDoctor.setId(doctor.getId());
-
-        repository.save(newDoctor);
+        return this.repository.findAll()
+                .stream()
+                .map(mapToResponseDto())
+                .collect(Collectors.toList());
     }
 
-    public void delete(UUID uuid){
-      repository.deleteByUuid(uuid);
+    public DoctorDtoResponse findByUuid(UUID uuid) {
+        return this.repository.findByUuid(uuid)
+                .map(mapToResponseDto())
+                .orElseThrow(DoctorNotFoundException::new);
     }
 
+    public DoctorDtoResponse addDoctor(DoctorDtoRequest dto) {
+        Doctor doctor = this.modelMapper.map(dto, Doctor.class);
+        return modelMapper.map(repository.save(doctor), DoctorDtoResponse.class);
+    }
+
+    public void update(UUID uuid, DoctorDtoRequest newDoctor) {
+        Doctor doctor = repository.findByUuid(uuid).orElseThrow(DoctorNotFoundException::new);
+        doctor.setName(newDoctor.getName());
+        doctor.setCrm(newDoctor.getCrm());
+        doctor.setSpecialty(newDoctor.getSpecialty());
+        repository.save(doctor);
+    }
+
+    @Transactional
+    public void delete(UUID uuid) {
+        if (!repository.existsByUuid(uuid)) {
+            throw new DoctorNotFoundException();
+        }
+        repository.deleteByUuid(uuid);
+    }
+
+    private Function<Doctor, DoctorDtoResponse> mapToResponseDto() {
+        return doctor -> modelMapper.map(doctor, DoctorDtoResponse.class);
+    }
 }
